requirements.txt
hyperliquid-python-sdk~=0.3.0
eth-account~=0.12.1
